Fix Fast Refresh lint warnings
