+++ apps/web/src/components/app/activity-chart-utils.ts
@@ -1,0 +1,5 @@
+import { formatShortDate } from "@/lib/format";
+
+export function formatDate(iso: string): string {
+  return formatShortDate(iso);
+}
