+++ apps/web/src/components/app/agent-history-event-types.ts
@@ -1,0 +1,15 @@
+export const EVENT_TYPE_COLORS: Record<string, string> = {
+  working: "bg-status-working",
+  blocked: "bg-status-blocked",
+  waiting_user: "bg-status-waiting",
+  done: "bg-status-done",
+  idle: "bg-muted-foreground",
+};
+
+export const EVENT_TYPE_LABELS: Record<string, string> = {
+  working: "Working",
+  blocked: "Blocked",
+  waiting_user: "Waiting",
+  done: "Done",
+  idle: "Idle",
+};
