+++ apps/web/src/components/app/feedback-mobile.tsx
@@ -23,7 +23,7 @@ import {
   getVerdict,
   getReviewSummary,
   getFilesReviewed,
-} from "@/components/app/persona-agent-row";
+} from "@/components/app/persona-agent-review-utils";
 import { type Agent, type FeedbackItem } from "@/components/app/types";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
