+++ apps/web/src/components/app/persona-agent-row.tsx
@@ -12,6 +12,11 @@ import {
   reviewVerdictLabel,
   type ReviewVerdict,
 } from "@/components/app/agent-event-utils";
+import {
+  getResolution,
+  getReviewSummary,
+  getVerdict,
+} from "@/components/app/persona-agent-review-utils";
 import { type Agent, type AgentVisualState } from "@/components/app/types";
 import { Badge } from "@/components/ui/badge";
 import {
@@ -144,25 +149,6 @@ function PersonaVerdictIcon({
       <Flag className="h-3 w-3" />
     </span>
   );
-}
-
-export function getVerdict(child: Agent): ReviewVerdict | undefined {
-  const v = child.review?.verdict;
-  if (v === "approve" || v === "request_changes") return v;
-  return undefined;
-}
-
-export function getReviewSummary(child: Agent): string | undefined {
-  return child.review?.summary ?? undefined;
-}
-
-export function getFilesReviewed(child: Agent): string[] | undefined {
-  const f = child.review?.filesReviewed;
-  return Array.isArray(f) ? f : undefined;
-}
-
-export function getResolution(child: Agent) {
-  return child.review?.resolution ?? undefined;
 }
 
 type StepColor = "orange" | "emerald" | "muted";
