+++ apps/web/src/components/app/settings-pane.tsx
@@ -1,14 +1,4 @@
-import { useEffect, useState } from "react";
-import {
-  ArrowDownToLine,
-  Bell,
-  BookOpenText,
-  Database,
-  Package,
-  Server,
-  Settings,
-  Users,
-} from "lucide-react";
+import { Database, Server, Settings } from "lucide-react";
 
 import { AgentTypeSettings } from "@/components/app/agent-type-settings";
 import { AppearanceSettings } from "@/components/app/appearance-settings";
@@ -28,53 +18,8 @@ import { useReleaseStream } from "@/hooks/use-release-stream";
 import { type ThemeId } from "@/hooks/use-theme";
 import { type AgentType } from "@/lib/agent-types";
 import { type IdeType } from "@/lib/ide-types";
-import { api } from "@/lib/api";
+import { type SettingsSection } from "@/components/app/settings-state";
 import { cn } from "@/lib/utils";
-
-type SettingsSection =
-  | "general"
-  | "agents"
-  | "notifications"
-  | "updates"
-  | "help"
-  | "releases";
-
-const BASE_SECTIONS: Array<{
-  id: SettingsSection;
-  label: string;
-  icon: typeof ArrowDownToLine;
-}> = [
-  { id: "general", label: "General", icon: Settings },
-  { id: "agents", label: "Agents", icon: Users },
-  { id: "notifications", label: "Notifications", icon: Bell },
-  { id: "updates", label: "Updates", icon: ArrowDownToLine },
-];
-
-const RELEASES_SECTION = {
-  id: "releases" as SettingsSection,
-  label: "Releases",
-  icon: Package,
-};
-const HELP_SECTION = {
-  id: "help" as SettingsSection,
-  label: "Help",
-  icon: BookOpenText,
-};
-
-const ALL_VALID_SECTIONS: SettingsSection[] = [
-  "general",
-  "agents",
-  "notifications",
-  "updates",
-  "help",
-  "releases",
-];
-
-function isValidSection(value: string | undefined): value is SettingsSection {
-  return (
-    value !== undefined && ALL_VALID_SECTIONS.includes(value as SettingsSection)
-  );
-}
 
 export type SettingsPaneProps = {
   open: boolean;
@@ -98,44 +43,6 @@ export type SettingsPaneProps = {
   onSectionChange?: (section: string | null) => void;
   onSubsectionChange?: (subsection: string | null) => void;
 };
-
-export function useSettingsState(open: boolean, initialSection?: string) {
-  const resolvedInitial = isValidSection(initialSection)
-    ? initialSection
-    : "general";
-  const [activeSection, setActiveSectionState] =
-    useState<SettingsSection | null>(resolvedInitial);
-  const [isAdmin, setIsAdmin] = useState(false);
-
-  useEffect(() => {
-    if (!open) return;
-    let cancelled = false;
-    void api<{ isAdmin: boolean }>("/api/v1/release/admin-check")
-      .then((data) => {
-        if (!cancelled) setIsAdmin(data.isAdmin);
-      })
-      .catch(() => {});
-    return () => {
-      cancelled = true;
-    };
-  }, [open]);
-
-  const sections = isAdmin
-    ? [...BASE_SECTIONS, RELEASES_SECTION, HELP_SECTION]
-    : [...BASE_SECTIONS, HELP_SECTION];
-
-  useEffect(() => {
-    if (open && isValidSection(initialSection)) {
-      if (initialSection === "releases" && !isAdmin) {
-        setActiveSectionState("general");
-      } else {
-        setActiveSectionState(initialSection);
-      }
-    }
-  }, [open, initialSection, isAdmin]);
-
-  return { activeSection, setActiveSectionState, isAdmin, sections };
-}
 
 /** Settings nav for the sidebar. */
 export function SettingsNavContent({
