+++ apps/web/src/contexts/auth-context-core.ts
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+
+import { type AuthState } from "@/components/app/types";
+
+export type AuthContextValue = {
+  authState: AuthState;
+  handleAuthenticated: () => void;
+  handleLogout: () => Promise<void>;
+  retryAuth: () => void;
+};
+
+export const AuthContext = createContext<AuthContextValue | null>(null);
