+++ apps/web/src/contexts/auth-context-provider.tsx
@@ -1,0 +1,14 @@
+import {
+  AuthContext,
+  type AuthContextValue,
+} from "@/contexts/auth-context-core";
+
+export function AuthContextProvider({
+  children,
+  value,
+}: {
+  children: React.ReactNode;
+  value: AuthContextValue;
+}): JSX.Element {
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+}
