+++ apps/web/src/contexts/auth-context.tsx
@@ -1,16 +1,9 @@
-import { createContext, useContext } from "react";
-import { type AuthState } from "@/components/app/types";
+import { useContext } from "react";
 
-type AuthContextValue = {
-  authState: AuthState;
-  handleAuthenticated: () => void;
-  handleLogout: () => Promise<void>;
-  retryAuth: () => void;
-};
-
-const AuthContext = createContext<AuthContextValue | null>(null);
-
-export const AuthContextProvider = AuthContext.Provider;
+import {
+  AuthContext,
+  type AuthContextValue,
+} from "@/contexts/auth-context-core";
 
 export function useAuthContext(): AuthContextValue {
   const ctx = useContext(AuthContext);
